requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
anthropic>=0.25.0
google-genai>=1.0.0
gspread>=5.0.0
google-auth>=2.0.0


